Bugfix: button image needs global reference
The image button did not work since the image reference was passed as an argument to the tk.Button object, but not stored internally. Since the button and with it the image holding variable is created in a function, and the tk.Button object does not keep a reference to the image internally, the only image reference is the variable in that function. When the function call ends and the button is defined, so to was the image reference deleted again by the garbage collector.