Decoupled view & cntrllr, impl. threading & deflts
Decoupled the view and the controller by making use of a protocol as abstract interface inbetween the two. This also fixed the problem with circular dependencies. Furthermore threading was implemented to allow for asynchronous updates to the GUI, while processing is running in the background. Finally also finished and tested implementation of defaults via JSON-file and finalized error handling.